+++ apptest.go
@@ -5,16 +5,23 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/giantswarm/apiextensions/v2/pkg/apis/application/v1alpha1"
-	"github.com/giantswarm/apiextensions/v2/pkg/label"
+	v1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/application/v1alpha1"
+	"github.com/giantswarm/apiextensions/v3/pkg/label"
 	"github.com/giantswarm/appcatalog"
 	"github.com/giantswarm/backoff"
-	"github.com/giantswarm/k8sclient/v4/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
 )
 
 const (
@@ -25,28 +32,83 @@ const (
 
 // Config represents the configuration used to setup the apps.
 type Config struct {
-	K8sClient k8sclient.Interface
-	Logger    micrologger.Logger
+	KubeConfigPath string
+
+	Logger micrologger.Logger
 }
 
 // AppSetup implements the logic for managing the app setup.
 type AppSetup struct {
-	k8sClient k8sclient.Interface
-	logger    micrologger.Logger
+	ctrlClient client.Client
+	k8sClient  kubernetes.Interface
+	logger     micrologger.Logger
 }
 
 // New creates a new configured app setup library.
 func New(config Config) (*AppSetup, error) {
+	var err error
+
+	if config.KubeConfigPath == "" {
+		return nil, microerror.Maskf(invalidConfigError, "%T.KubeConfig must not be empty", config)
+	}
+
 	if config.Logger == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
 	}
-	if config.K8sClient == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.K8sClient must not be empty", config)
+
+	var restConfig *rest.Config
+	{
+		restConfig, err = clientcmd.BuildConfigFromFlags("", config.KubeConfigPath)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
+	var ctrlClient client.Client
+	{
+		// Extend the global client-go scheme which is used by all the tools under
+		// the hood. The scheme is required for the controller-runtime controller to
+		// be able to watch for runtime objects of a certain type.
+		appSchemeBuilder := runtime.SchemeBuilder(schemeBuilder{
+			v1alpha1.AddToScheme,
+		})
+		err = appSchemeBuilder.AddToScheme(scheme.Scheme)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+
+		// Configure a dynamic rest mapper to the controller client so it can work
+		// with runtime objects of arbitrary types. Note that this is the default
+		// for controller clients created by controller-runtime managers.
+		// Anticipating a rather uncertain future and more breaking changes to come
+		// we want to separate client and manager. Thus we configure the client here
+		// properly on our own instead of relying on the manager to provide a
+		// client, which might change in the future.
+		mapper, err := apiutil.NewDynamicRESTMapper(rest.CopyConfig(restConfig))
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+
+		ctrlClient, err = client.New(rest.CopyConfig(restConfig), client.Options{Scheme: scheme.Scheme, Mapper: mapper})
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
+	var k8sClient kubernetes.Interface
+	{
+		c := rest.CopyConfig(restConfig)
+
+		k8sClient, err = kubernetes.NewForConfig(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
 	}
 
 	a := &AppSetup{
-		k8sClient: config.K8sClient,
-		logger:    config.Logger,
+		ctrlClient: ctrlClient,
+		k8sClient:  k8sClient,
+		logger:     config.Logger,
 	}
 
 	return a, nil
@@ -75,6 +137,11 @@ func (a *AppSetup) InstallApps(ctx context.Context, apps []App) error {
 	return nil
 }
 
+// K8sClient returns a Kubernetes clienset for use in automated tests.
+func (a *AppSetup) K8sClient() kubernetes.Interface {
+	return a.k8sClient
+}
+
 func (a *AppSetup) createAppCatalogs(ctx context.Context, apps []App) error {
 	var err error
 
@@ -98,7 +165,7 @@ func (a *AppSetup) createAppCatalogs(ctx context.Context, apps []App) error {
 				},
 			},
 		}
-		_, err = a.k8sClient.G8sClient().ApplicationV1alpha1().AppCatalogs().Create(ctx, appCatalogCR, metav1.CreateOptions{})
+		err = a.ctrlClient.Create(ctx, appCatalogCR)
 		if apierrors.IsAlreadyExists(err) {
 			a.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("%#q appcatalog CR already exists", appCatalogCR.Name))
 		} else if err != nil {
@@ -158,7 +225,7 @@ func (a *AppSetup) createApps(ctx context.Context, apps []App) error {
 			appCR.Spec.UserConfig.ConfigMap.Namespace = namespace
 		}
 
-		_, err = a.k8sClient.G8sClient().ApplicationV1alpha1().Apps(namespace).Create(ctx, appCR, metav1.CreateOptions{})
+		err = a.ctrlClient.Create(ctx, appCR)
 		if apierrors.IsAlreadyExists(err) {
 			a.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("%#q app CR already exists", appCR.Name))
 			return nil
@@ -186,7 +253,7 @@ func (a *AppSetup) createUserValuesConfigMap(ctx context.Context, name, namespac
 		},
 	}
 
-	_, err := a.k8sClient.K8sClient().CoreV1().ConfigMaps(namespace).Create(ctx, configMap, metav1.CreateOptions{})
+	_, err := a.k8sClient.CoreV1().ConfigMaps(namespace).Create(ctx, configMap, metav1.CreateOptions{})
 	if apierrors.IsAlreadyExists(err) {
 		a.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("already created configmap %#q", name))
 	} else if err != nil {
@@ -214,16 +281,24 @@ func (a *AppSetup) waitForDeployedApps(ctx context.Context, apps []App) error {
 }
 
 func (a *AppSetup) waitForDeployedApp(ctx context.Context, appName string) error {
+	var err error
+
 	a.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("ensuring %#q app CR is %#q", appName, deployedStatus))
 
+	var app v1alpha1.App
+
 	o := func() error {
-		app, err := a.k8sClient.G8sClient().ApplicationV1alpha1().Apps(namespace).Get(ctx, appName, metav1.GetOptions{})
+		err = a.ctrlClient.Get(
+			ctx,
+			types.NamespacedName{Name: appName, Namespace: namespace},
+			&app)
 		if err != nil {
 			return microerror.Mask(err)
 		}
 		if app.Status.Release.Status != deployedStatus {
 			return microerror.Maskf(executionFailedError, "waiting for %#q, current %#q", deployedStatus, app.Status.Release.Status)
 		}
+
 		return nil
 	}
 
@@ -231,8 +306,8 @@ func (a *AppSetup) waitForDeployedApp(ctx context.Context, appName string) error
 		a.logger.Log("level", "debug", "message", fmt.Sprintf("failed to get app CR status '%s': retrying in %s", deployedStatus, t), "stack", fmt.Sprintf("%v", err))
 	}
 
-	b := backoff.NewConstant(20*time.Minute, 60*time.Second)
-	err := backoff.RetryNotify(o, b, n)
+	b := backoff.NewConstant(20*time.Minute, 30*time.Second)
+	err = backoff.RetryNotify(o, b, n)
 	if err != nil {
 		return microerror.Mask(err)
 	}