+++ spec.go
@@ -1,11 +1,19 @@
 package apptest
 
-import "context"
+import (
+	"context"
+
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/client-go/kubernetes"
+)
 
 type Interface interface {
 	// InstallApps creates appcatalog and app CRs for use in automated tests
 	// and ensures they are installed by our app platform.
 	InstallApps(ctx context.Context, apps []App) error
+
+	// K8sClient returns a Kubernetes clienset for use in automated tests.
+	K8sClient() kubernetes.Interface
 }
 
 type App struct {
@@ -18,3 +26,6 @@ type App struct {
 	Version       string
 	WaitForDeploy bool
 }
+
+// schemeBuilder is used to extend the known types of the client-go scheme.
+type schemeBuilder []func(*runtime.Scheme) error